// front-office/src/components/atoms/SEO/type.ts
export enum MetaKeys {
  TITLE = 'title',
  DESCRIPTION = 'description',
  CANONICAL = 'canonical',
  KEYWORDS = 'keywords',

  OG_IMAGE = 'ogImage',
  OG_URL = 'ogUrl',
  OG_TYPE = 'ogType',
  OG_SITE_NAME = 'ogSiteName',
  OG_DESCRIPTION = 'ogDescription',
  OG_TITLE = 'ogTitle',
  OG_PUBLISH_DATE = 'ogPublishDate',

  TWITTER_CARD = 'twitterCard',
  TWITTER_SITE = 'twitterSite',
  TWITTER_TITLE = 'twitterTitle',
  TWITTER_DESCRIPTION = 'twitterDescription',
  TWITTER_IMAGE = 'twitterImage',

  ARTICLE_PUBLISHED_TIME = 'articlePublishedTime',
  ARTICLE_AUTHOR = 'articleAuthor',

  ROBOTS = 'robots'
}